Share conversation list rendering between screens
